+++ wled00/fcn_declare.h
@@ -13,13 +13,6 @@ void handleAlexa();
 void onAlexaChange(EspalexaDevice* dev);
 #endif
 
-//blynk.cpp
-#ifndef WLED_DISABLE_BLYNK
-void initBlynk(const char* auth, const char* host, uint16_t port);
-void handleBlynk();
-void updateBlynk();
-#endif
-
 //button.cpp
 void shortPressAction(uint8_t b=0);
 void longPressAction(uint8_t b=0);